requirements2.txt
-e tabarena/[benchmark]
autogluon==1.4.1b20251023          
autogluon.tabular==1.4.1b20251023
autogluon.core==1.4.1b20251023
autogluon.features==1.4.1b20251023
torch==2.7.1 --index-url https://download.pytorch.org/whl/cu126
torchvision==0.22.1
triton==3.3.1
numpy==2.1.3
pandas==2.3.3
scikit-learn==1.6.1
scipy==1.16.3
pyarrow==20.0.0        
lightgbm==4.6.0
xgboost==3.0.5
catboost==1.2.8
tabpfn==2.2.1
tabpfn-common-utils==0.2.4
tabdpt==1.1.7               
tabicl==0.1.3              
ray==2.48.0               
transformers==4.49.0
huggingface-hub==0.36.0
tokenizers==0.21.4
safetensors==0.6.2
accelerate==1.11.0
datasets==4.0.0
evaluate==0.4.6
tqdm==4.67.1
joblib==1.5.2
threadpoolctl==3.6.0      
psutil==7.0.0             
optuna==4.5.0              
ConfigSpace==1.2.1
matplotlib==3.10.1
seaborn==0.13.2             
plotly==6.3.1               
pydantic==2.12.3         
pyyaml==6.0.3
